[outil-csv] refacto: extraction d'une fonction subscriberToReceipient
